+++ src/crd/i18n/dashboard/dashboard.bg.json
@@ -18,7 +18,9 @@
     "homeSpacePlaceholder": {
       "title": "Задай начално пространство",
       "subtitle": "Закачи любимото си пространство на таблото"
-    }
+    },
+    "homeSpaceSettings": "Настройки на началното пространство",
+    "private": "Частно"
   },
   "activity": {
     "spacesTitle": "Последна активност в моите пространства",
@@ -30,7 +32,10 @@
       },
       "role": {
         "label": "Моята роля",
-        "all": "Всички роли"
+        "all": "Всички роли",
+        "member": "Член",
+        "lead": "Лидер",
+        "admin": "Администратор"
       }
     },
     "showMore": "Покажи повече",
